Move `createStructuredSelector` type tests into its own file
